requirements.txt
numpy
flax
orbax
optax
optuna
hydra-core
wandb
matplotlib

# shapenet
trimesh

--extra-index-url https://storage.googleapis.com/jax-releases/jax_cuda_releases.htm
jax[cuda12_pip]

--extra-index-url https://download.pytorch.org/whl/cpu
torch 
torchvision 